Be noisy about updating permissions
In #75, I mentioned that I was a little stumped why my Nextcloud container was showing a `502 Bad Gateway` error. In the end, it turned out that it was just updating the permissions in the container, which are set to `nobody:nogroup`, but need to be chowned to `991:991` when first starting a fresh container. This PR seeks to make that a little more obvious to the user, in case they choose to use `docker-compose logs -f nextcloud` or such.